% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_cellclassify_garnett_function.r
\name{classify_pbmc_cells}
\alias{classify_pbmc_cells}
\title{Classify PBMC SC data}
\usage{
classify_pbmc_cells(
  seurat_data,
  path_10x_data,
  sample_names = FALSE,
  marker_file_path,
  lower_gene_filter_num = 200,
  pbmc_classifier = FALSE,
  cds_gene_type = "ENSEMBL",
  downsampled = TRUE,
  save_object = TRUE,
  results_folder = "s2_garnett_classifier_results"
)
}
\arguments{
\item{seurat_data}{filtered data from seurat}

\item{path_10x_data}{path to 10x folder}

\item{marker_file_path}{path to the marker file needed for garnett classification}

\item{lower_gene_filter_num}{minimum number of genes that has to be detected in a cell}

\item{pbmc_classifier}{path to pre built classifier if one is being used (defualt is to build its own classifier)}

\item{cds_gene_type}{type of gene ID (defualt is Ensembl)}

\item{save_object}{save seurat normalized object in RDS file so this step doesn't have to be redone everytime (defualt is TRUE)}

\item{results_folder}{path to results of analysis (defualt is s2_garnett_classifier_results)}
}
\description{
This generates new folder (internal function only)
}
\keyword{cell}
\keyword{classification}
\keyword{classifier}
\keyword{garnett}
\keyword{type}
\keyword{using}
