% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_cellclassify_garnett.r
\name{final_cds_garnett_classification}
\alias{final_cds_garnett_classification}
\title{Classify PBMC SC data}
\usage{
final_cds_garnett_classification(cds_garnett)
}
\arguments{
\item{seurat_data}{filtered data from seurat}

\item{path_10x_data}{path to 10x folder}

\item{marker_file_path}{path to the marker file needed for garnett classification}

\item{pbmc_classifier}{path to pre built classifier if one is being used (defualt is to build its own classifier)}

\item{cds_gene_type}{type of gene ID (defualt is Ensembl)}

\item{results_folder}{path to results of analysis (defualt is s2_garnett_classifier_results)}
}
\description{
This generates new folder (internal function only)
}
\keyword{cell}
\keyword{classification}
\keyword{classifier}
\keyword{garnett}
\keyword{type}
\keyword{using}
