% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_folder_function.r
\name{generate_folder}
\alias{generate_folder}
\title{generate folder function}
\usage{
generate_folder(foldername)
}
\arguments{
\item{foldername}{folder to genrate}
}
\description{
This generates new folder (internal function only)
}
\keyword{remove}
\keyword{samples}
