% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_initial_load_data_function.r
\name{s1_load_data}
\alias{s1_load_data}
\title{Initially load data into Seurat}
\usage{
s1_load_data(
  path_10_data,
  sample_names,
  lower_gene_filter_num = 200,
  upper_gene_filter_num = 2500,
  percent_mito = 10,
  downsampled = TRUE,
  top_variable_genes = 10,
  run_umap = TRUE,
  run_tsne = FALSE,
  save_object = TRUE,
  results_folder = "s1_quality_control_results"
)
}
\arguments{
\item{lower_gene_filter_num}{minimum number of genes that has to be detected in a cell}

\item{upper_gene_filter_num}{maximum number of genes that can be detected in a cell}

\item{percent_mito}{allowable percenatage of mitochondria DNA allowed}

\item{downsampled}{if downsampled feature was used when cell ranger was run}

\item{top_variable_genes}{how many genes to label in variablity plots (defualt is 10)}

\item{run_umap}{run umap for feature reduction}

\item{run_tsne}{run tsne for feature reduction}

\item{save_object}{save seurat normalized object in RDS file so this step doesn't have to be redone everytime (defualt is TRUE)}

\item{results_folder}{folder that results appear in (default: s1_quality_control_results)}

\item{path_10x_data}{path to 10x folder}
}
\description{
This generates new folder (internal function only)
}
\keyword{data}
\keyword{load}
