% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_DE_function.r
\name{s3_DE_analysis}
\alias{s3_DE_analysis}
\title{perform differential gene expression}
\usage{
s3_DE_analysis(
  pbmc,
  ident_1,
  ident_2,
  LFC = 0.26,
  pvalue = 0.05,
  clusters_DE = FALSE,
  individual_gene_plots = FALSE,
  results_folder = "s3_DE_results"
)
}
\arguments{
\item{pbmc}{filtered data from seurat}

\item{ident_1}{cell condition 1}

\item{ident_2}{cell condition 2}

\item{LFC}{log fold cutoff (defualt 0.26)}

\item{pvalue}{pvalue (FDR cutoff (default 0.05)}

\item{clusters_DE}{perform DE on individual clusters in ident_1 & 2 (defualt is TRUE, needs s2_cellclasify_garnett to have been run)}

\item{individual_gene_plots}{generate individual violin plots for every DE gene (defualt is FALSE)}

\item{results_folder}{path to results of analysis (defualt is s2_garnett_classifier_results)}
}
\description{
This generates new folder (internal function only)
}
\keyword{analysis}
\keyword{differential}
\keyword{gene}
\keyword{perform}
