% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1_initial_load_data_function.r
\name{s1_load_data}
\alias{s1_load_data}
\title{Initially load data into Seurat}
\usage{
s1_load_data(
  path_10_data,
  sample_names,
  lower_gene_filter_num = 200,
  upper_gene_filter_num = 2500,
  percent_mito = 10,
  downsampled = TRUE,
  run_umap = TRUE,
  run_tsne = FALSE,
  results_folder = "s1_quality_control_results"
)
}
\arguments{
\item{lower_gene_filter_num}{minimum number of genes that has to be detected in a cell}

\item{upper_gene_filter_num}{maximum number of genes that can be detected in a cell}

\item{percent_mito}{allowable percenatage of mitochondria DNA allowed}

\item{downsampled}{if downsampled feature was used when cell ranger was run}

\item{run_umap}{run umap for feature reduction}

\item{run_tsne}{run tsne for feature reduction}

\item{results_folder}{folder that results appear in (default: s1_quality_control_results)}

\item{path_10x_data}{path to 10x folder}
}
\description{
This generates new folder (internal function only)
}
\keyword{data}
\keyword{load}
